requirements.txt
# Core dependencies
requests==2.31.0
feedparser==6.0.10
pydub==0.25.1
spacy==3.7.2
pytextrank==3.2.4
sentence-transformers==2.2.2
faiss-cpu==1.7.4
transformers==4.36.2
numpy==1.26.2
soundfile==0.12.1
torch==2.1.1
feedgen==0.9.0

# TTS dependency (specific version may depend on model compatibility)
tts==0.22.0

# Stable Diffusion dependency
diffusers==0.23.0

# spaCy language model (to be installed separately via spacy download)
# Run: python -m spacy download en_core_web_sm